feat: make the modal backwards compatible
