+++ src/support/slack/actions/onboard-modal.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Adobe. All rights reserved.
+ * Copyright 2025 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -10,15 +10,10 @@
  * governing permissions and limitations under the License.
  */
 
-import { Site as SiteModel, Organization as OrganizationModel } from '@adobe/spacecat-shared-data-access';
-import { Config } from '@adobe/spacecat-shared-data-access/src/models/site/config.js';
-import {
-  isValidUrl, isObject, resolveCanonicalUrl,
-} from '@adobe/spacecat-shared-utils';
-import { SFNClient, StartExecutionCommand } from '@aws-sdk/client-sfn';
-
-import { findDeliveryType, triggerImportRun, triggerAuditForSite } from '../../utils.js';
+import { onboardSingleSite as sharedOnboardSingleSite } from '../../utils.js';
 import { loadProfileConfig } from '../../../utils/slack/base.js';
+
+export const AEM_CS_HOST = /^author-p(\d+)-e(\d+)/i;
 
 /**
  * Extracts program and environment ID from AEM Cloud Service preview URLs.
@@ -30,15 +25,12 @@ function extractDeliveryConfigFromPreviewUrl(previewUrl) {
     const url = new URL(previewUrl);
     const { hostname } = url;
 
-    // Pattern for AEM Cloud Service URLs
-    const aemcsPattern = /(?:(?:author|publish|preview)-)?p(\d+)-e(\d+)\.(?:live\.)?adobeaemcloud\.com$/i;
-    const match = hostname.match(aemcsPattern);
+    const [, programId, envId] = AEM_CS_HOST.exec(hostname);
 
-    if (match) {
-      const [, programId, environmentId] = match;
+    if (programId && envId) {
       return {
-        programId,
-        environmentId,
+        programId: `p${programId}`,
+        environmentId: `e${envId}`,
         authorURL: previewUrl,
       };
     }
@@ -50,44 +42,7 @@ function extractDeliveryConfigFromPreviewUrl(previewUrl) {
 }
 
 /**
- * Checks if a site is in the LA_CUSTOMERS list and returns appropriate response if it is.
- */
-const checkLACustomerRestriction = async (
-  baseURL,
-  imsOrgID,
-  profileName,
-  slackContext,
-  env,
-  log,
-) => {
-  const { say } = slackContext;
-
-  if (env.LA_CUSTOMERS) {
-    const laCustomers = env.LA_CUSTOMERS.split(',').map((url) => url.trim());
-    const isLACustomer = laCustomers.some(
-      (url) => baseURL.toLowerCase().endsWith(url.toLowerCase()),
-    );
-
-    if (isLACustomer) {
-      const message = `:warning: Cannot onboard site ${baseURL} - it's already onboarded and live!`;
-      log.warn(message);
-      await say(message);
-      return {
-        site: baseURL,
-        imsOrgId: imsOrgID,
-        profile: profileName,
-        errors: 'Site is a Live customer',
-        status: 'Failed',
-        existingSite: 'Yes',
-      };
-    }
-  }
-
-  return null;
-};
-
-/**
- * Onboards a single site from modal input.
+ * Onboards a single site from modal input
  */
 const onboardSingleSiteFromModal = async (
   baseURLInput,
@@ -99,334 +54,21 @@ const onboardSingleSiteFromModal = async (
   context,
   additionalParams = {},
 ) => {
-  const { say } = slackContext;
-  const {
-    dataAccess, log, imsClient, env,
-  } = context;
-  const { Site, Organization } = dataAccess;
-  const sfnClient = new SFNClient();
+  // Load the profile configuration
+  const profile = await loadProfileConfig(profileName);
 
-  const baseURL = baseURLInput.trim();
-  const imsOrgID = imsOrganizationID || env.DEMO_IMS_ORG;
-
-  // Check if site is in LA_CUSTOMERS list
-  const laCustomerCheck = await checkLACustomerRestriction(
-    baseURL,
-    imsOrgID,
-    profileName,
+  // Use the shared onboarding function from utils
+  return sharedOnboardSingleSite(
+    baseURLInput,
+    imsOrganizationID,
+    configuration,
+    profile,
+    workflowWaitTime,
     slackContext,
-    env,
-    log,
+    context,
+    additionalParams,
+    { profileName },
   );
-  if (laCustomerCheck) {
-    return laCustomerCheck;
-  }
-
-  log.info(`Starting ${profileName} environment setup for site ${baseURL}`);
-  await say(`:gear: Starting ${profileName} environment setup for site ${baseURL}`);
-  await say(':key: Please make sure you have access to the AEM Shared Production Demo environment. Request access here: https://demo.adobe.com/demos/internal/AemSharedProdEnv.html');
-
-  const reportLine = {
-    site: baseURL,
-    imsOrgId: imsOrgID,
-    spacecatOrgId: '',
-    siteId: '',
-    profile: profileName,
-    deliveryType: '',
-    authoringType: additionalParams.authoringType || '',
-    imports: '',
-    audits: '',
-    errors: '',
-    status: 'Success',
-    existingSite: 'No',
-  };
-
-  try {
-    if (!isValidUrl(baseURL)) {
-      reportLine.errors = 'Invalid site base URL';
-      reportLine.status = 'Failed';
-      return reportLine;
-    }
-
-    if (!OrganizationModel.IMS_ORG_ID_REGEX.test(imsOrgID)) {
-      reportLine.errors = 'Invalid IMS Org ID';
-      reportLine.status = 'Failed';
-      return reportLine;
-    }
-
-    // check if the organization with IMS Org ID already exists; create if it doesn't
-    let organization = await Organization.findByImsOrgId(imsOrgID);
-    if (!organization) {
-      let imsOrgDetails;
-      try {
-        imsOrgDetails = await imsClient.getImsOrganizationDetails(imsOrgID);
-        log.info(`IMS Org Details: ${imsOrgDetails}`);
-      } catch (error) {
-        log.error(`Error retrieving IMS Org details: ${error.message}`);
-        reportLine.errors = `Error retrieving IMS org with the ID *${imsOrgID}*.`;
-        reportLine.status = 'Failed';
-        return reportLine;
-      }
-
-      if (!imsOrgDetails) {
-        reportLine.errors = `Could not find details of IMS org with the ID *${imsOrgID}*.`;
-        reportLine.status = 'Failed';
-        return reportLine;
-      }
-
-      organization = await Organization.create({
-        name: imsOrgDetails.orgName,
-        imsOrgId: imsOrgID,
-      });
-
-      const message = `:white_check_mark: A new organization has been created. Organization ID: ${organization.getId()} Organization name: ${organization.getName()} IMS Org ID: ${imsOrgID}.`;
-      await say(message);
-      log.info(message);
-    }
-
-    const organizationId = organization.getId();
-    log.info(`Organization ${organizationId} was successfully retrieved or created`);
-    reportLine.spacecatOrgId = organizationId;
-
-    let site = await Site.findByBaseURL(baseURL);
-    if (site) {
-      reportLine.existingSite = 'Yes';
-      reportLine.deliveryType = site.getDeliveryType();
-      log.info(`Site ${baseURL} already exists. Site ID: ${site.getId()}, Delivery Type: ${reportLine.deliveryType}`);
-
-      const siteOrgId = site.getOrganizationId();
-      if (siteOrgId !== organizationId) {
-        log.info(`:warning: :alert: Site ${baseURL} organization ID mismatch. Run below slack command to update site organization to ${organizationId}`);
-        log.info(`:fire: @spacecat set imsorg ${baseURL} ${organizationId}`);
-      }
-    } else {
-      log.info(`Site ${baseURL} doesn't exist. Finding delivery type...`);
-
-      // Use forced delivery type if provided, otherwise detect it
-      let deliveryType;
-      if (additionalParams.deliveryType
-        && Object.values(SiteModel.DELIVERY_TYPES).includes(additionalParams.deliveryType)) {
-        deliveryType = additionalParams.deliveryType;
-        log.info(`Using forced delivery type for site ${baseURL}: ${deliveryType}`);
-      } else {
-        deliveryType = await findDeliveryType(baseURL);
-        log.info(`Found delivery type for site ${baseURL}: ${deliveryType}`);
-      }
-
-      reportLine.deliveryType = deliveryType;
-      const isLive = deliveryType === SiteModel.DELIVERY_TYPES.AEM_EDGE;
-
-      const siteCreateParams = {
-        baseURL,
-        deliveryType,
-        isLive,
-        organizationId,
-      };
-
-      // Add authoring type if provided
-      if (additionalParams.authoringType
-        && Object.values(SiteModel.AUTHORING_TYPES || {})
-          .includes(additionalParams.authoringType)) {
-        siteCreateParams.authoringType = additionalParams.authoringType;
-        log.info(`Setting authoring type for site ${baseURL}: ${additionalParams.authoringType}`);
-      }
-
-      try {
-        site = await Site.create(siteCreateParams);
-      } catch (error) {
-        log.error(`Error creating site: ${error.message}`);
-        reportLine.errors = error.message;
-        reportLine.status = 'Failed';
-        return reportLine;
-      }
-    }
-
-    const siteID = site.getId();
-    log.info(`Site ${baseURL} was successfully retrieved or created. Site ID: ${siteID}`);
-    reportLine.siteId = siteID;
-
-    const profile = await loadProfileConfig(profileName);
-    log.info(`Profile ${profileName} was successfully loaded`);
-
-    if (!isObject(profile)) {
-      const error = `Profile "${profileName}" not found or invalid.`;
-      log.error(error);
-      reportLine.errors = error;
-      reportLine.status = 'Failed';
-      return reportLine;
-    }
-
-    if (!isObject(profile?.audits)) {
-      const error = `Profile "${profileName}" does not have a valid audits section.`;
-      log.error(error);
-      reportLine.errors = error;
-      reportLine.status = 'Failed';
-      return reportLine;
-    }
-
-    if (!isObject(profile?.imports)) {
-      const error = `Profile "${profileName}" does not have a valid imports section.`;
-      log.error(error);
-      reportLine.errors = error;
-      reportLine.status = 'Failed';
-      return reportLine;
-    }
-
-    const importTypes = Object.keys(profile.imports);
-    reportLine.imports = importTypes.join(', ');
-    const siteConfig = site.getConfig();
-    for (const importType of importTypes) {
-      siteConfig.enableImport(importType);
-    }
-
-    log.info(`Enabled the following imports for ${siteID}: ${reportLine.imports}`);
-
-    // Resolve canonical URL for the site from the base URL
-    const resolvedUrl = await resolveCanonicalUrl(baseURL);
-    const { pathname: baseUrlPathName } = new URL(baseURL);
-    const { pathname: resolvedUrlPathName, origin: resolvedUrlOrigin } = new URL(resolvedUrl);
-
-    log.info(`Base url: ${baseURL} -> Resolved url: ${resolvedUrl} for site ${siteID}`);
-
-    // Update the fetch configuration only if the pathname is different from the resolved URL
-    if (baseUrlPathName !== resolvedUrlPathName) {
-      siteConfig.updateFetchConfig({
-        overrideBaseURL: resolvedUrlOrigin,
-      });
-    }
-
-    site.setConfig(Config.toDynamoItem(siteConfig));
-    try {
-      await site.save();
-    } catch (error) {
-      log.error(error);
-      reportLine.errors = error.message;
-      reportLine.status = 'Failed';
-      return reportLine;
-    }
-
-    log.info(`Site config successfully saved for site ${siteID}`);
-
-    for (const importType of importTypes) {
-      /* eslint-disable no-await-in-loop */
-      await triggerImportRun(
-        configuration,
-        importType,
-        siteID,
-        profile.imports[importType].startDate,
-        profile.imports[importType].endDate,
-        slackContext,
-        context,
-      );
-    }
-
-    log.info(`Triggered the following imports for site ${siteID}: ${reportLine.imports}`);
-
-    const auditTypes = Object.keys(profile.audits);
-
-    auditTypes.forEach((auditType) => {
-      configuration.enableHandlerForSite(auditType, site);
-    });
-
-    reportLine.audits = auditTypes.join(', ');
-    log.info(`Enabled the following audits for site ${siteID}: ${reportLine.audits}`);
-
-    await say(`:white_check_mark: *Enabled imports*: ${reportLine.imports} *and audits*: ${reportLine.audits}`);
-
-    // trigger audit runs
-    log.info(`Starting audits for site ${baseURL}. Audit list: ${auditTypes}`);
-    await say(`:gear: Starting audits: ${auditTypes}`);
-    for (const auditType of auditTypes) {
-      /* eslint-disable no-await-in-loop */
-      if (!configuration.isHandlerEnabledForSite(auditType, site)) {
-        await say(`:x: Will not audit site '${baseURL}' because audits of type '${auditType}' are disabled for this site.`);
-      } else {
-        await triggerAuditForSite(
-          site,
-          auditType,
-          slackContext,
-          context,
-        );
-      }
-    }
-
-    // Opportunity status job
-    const opportunityStatusJob = {
-      type: 'opportunity-status-processor',
-      siteId: siteID,
-      siteUrl: baseURL,
-      imsOrgId: imsOrgID,
-      organizationId,
-      taskContext: {
-        auditTypes,
-        slackContext: {
-          channelId: slackContext.channelId,
-          threadTs: slackContext.threadTs,
-        },
-      },
-    };
-
-    // Disable imports and audits job
-    const disableImportAndAuditJob = {
-      type: 'disable-import-audit-processor',
-      siteId: siteID,
-      siteUrl: baseURL,
-      imsOrgId: imsOrgID,
-      organizationId,
-      taskContext: {
-        importTypes,
-        auditTypes,
-        slackContext: {
-          channelId: slackContext.channelId,
-          threadTs: slackContext.threadTs,
-        },
-      },
-    };
-
-    // Demo URL job
-    const demoURLJob = {
-      type: 'demo-url-processor',
-      siteId: siteID,
-      siteUrl: baseURL,
-      imsOrgId: imsOrgID,
-      organizationId,
-      taskContext: {
-        experienceUrl: env.EXPERIENCE_URL || 'https://experience.adobe.com',
-        slackContext: {
-          channelId: slackContext.channelId,
-          threadTs: slackContext.threadTs,
-        },
-      },
-    };
-
-    log.info(`Opportunity status job: ${JSON.stringify(opportunityStatusJob)}`);
-    log.info(`Disable import and audit job: ${JSON.stringify(disableImportAndAuditJob)}`);
-    log.info(`Demo URL job: ${JSON.stringify(demoURLJob)}`);
-
-    // Prepare and start step function workflow with the necessary parameters
-    const workflowInput = {
-      opportunityStatusJob,
-      disableImportAndAuditJob,
-      demoURLJob,
-      workflowWaitTime: workflowWaitTime || env.WORKFLOW_WAIT_TIME_IN_SECONDS,
-    };
-
-    const workflowName = `onboard-${baseURL.replace(/[^a-zA-Z0-9]/g, '-')}-${Date.now()}`;
-
-    const startCommand = new StartExecutionCommand({
-      stateMachineArn: env.ONBOARD_WORKFLOW_STATE_MACHINE_ARN,
-      input: JSON.stringify(workflowInput),
-      name: workflowName,
-    });
-    await sfnClient.send(startCommand);
-  } catch (error) {
-    log.error(error);
-    reportLine.errors = error.message;
-    reportLine.status = 'Failed';
-    throw error;
-  }
-
-  return reportLine;
 };
 
 /**
@@ -442,6 +84,17 @@ export function startOnboarding(lambdaContext) {
       await ack();
 
       const { user } = body;
+
+      // Parse initial values from button value for backwards compatibility
+      let initialValues = {};
+      try {
+        if (body.actions?.[0]?.value && body.actions[0].value !== 'start_onboarding') {
+          initialValues = JSON.parse(body.actions[0].value);
+        }
+      } catch (error) {
+        log.warn('Failed to parse initial values from button:', error);
+        initialValues = {};
+      }
 
       // Update the original message to show user's choice
       await respond({
@@ -485,6 +138,7 @@ export function startOnboarding(lambdaContext) {
                   type: 'plain_text',
                   text: 'https://site.url',
                 },
+                ...(initialValues.site && { initial_value: initialValues.site }),
               },
               label: {
                 type: 'plain_text',
@@ -499,8 +153,9 @@ export function startOnboarding(lambdaContext) {
                 action_id: 'ims_org_id',
                 placeholder: {
                   type: 'plain_text',
-                  text: 'ABC123@AdobeOrg (optional)',
+                  text: 'ABC123@AdobeOrg (default: AEM Sites Engineering)',
                 },
+                ...(initialValues.imsOrgId && { initial_value: initialValues.imsOrgId }),
               },
               label: {
                 type: 'plain_text',
@@ -516,15 +171,29 @@ export function startOnboarding(lambdaContext) {
                 action_id: 'profile',
                 placeholder: {
                   type: 'plain_text',
-                  text: 'Select profile',
+                  text: 'Select onboarding profile',
                 },
-                initial_option: {
-                  text: {
-                    type: 'plain_text',
-                    text: 'Default',
-                  },
-                  value: 'default',
-                },
+                initial_option: (() => {
+                  const profileOptions = [
+                    { text: 'Default', value: 'default' },
+                    { text: 'Demo', value: 'demo' },
+                    { text: 'Summit', value: 'summit' },
+                    { text: 'Summit – Lower Quality', value: 'summit-lower-quality' },
+                  ];
+
+                  const selectedProfile = initialValues.profile || 'default';
+                  const option = profileOptions.find(
+                    (opt) => opt.value === selectedProfile,
+                  ) || profileOptions[0];
+
+                  return {
+                    text: {
+                      type: 'plain_text',
+                      text: option.text,
+                    },
+                    value: option.value,
+                  };
+                })(),
                 options: [
                   {
                     text: {
@@ -596,7 +265,7 @@ export function startOnboarding(lambdaContext) {
                   {
                     text: {
                       type: 'plain_text',
-                      text: 'AEM Cloud Service',
+                      text: 'Adobe Managed Services',
                     },
                     value: 'aem_ams',
                   },
@@ -623,13 +292,13 @@ export function startOnboarding(lambdaContext) {
                 action_id: 'authoring_type',
                 placeholder: {
                   type: 'plain_text',
-                  text: 'Select authoring type (optional)',
+                  text: 'Select authoring type (required if preview URL is provided)',
                 },
                 options: [
                   {
                     text: {
                       type: 'plain_text',
-                      text: 'Document Authoring',
+                      text: 'Document Authoring (EDS or DA)',
                     },
                     value: 'documentauthoring',
                   },
@@ -643,7 +312,7 @@ export function startOnboarding(lambdaContext) {
                   {
                     text: {
                       type: 'plain_text',
-                      text: 'Crosswalk',
+                      text: 'Crosswalk (Universal Editor & EDS)',
                     },
                     value: 'cs/crosswalk',
                   },
@@ -668,6 +337,8 @@ export function startOnboarding(lambdaContext) {
                   type: 'plain_text',
                   text: '300 (default)',
                 },
+                ...(initialValues.workflowWaitTime
+                  && { initial_value: initialValues.workflowWaitTime.toString() }),
               },
               label: {
                 type: 'plain_text',
@@ -682,7 +353,7 @@ export function startOnboarding(lambdaContext) {
               type: 'section',
               text: {
                 type: 'mrkdwn',
-                text: '*Preview Environment Configuration* _(Optional)_\nConfigure preview environment for preflight and auto-fix.',
+                text: '*Preview Environment Configuration* _(Optional)_\nConfigure preview environment for preflight and auto-optimize. Only needed for AEM Cloud Service URLs.',
               },
             },
             {
@@ -761,6 +432,17 @@ export function onboardSiteModal(lambdaContext) {
           });
           return;
         }
+
+        // Require authoring type when preview URL is provided
+        if (!authoringType) {
+          await ack({
+            response_action: 'errors',
+            errors: {
+              authoring_type_input: 'Authoring type is required when a preview URL is provided.',
+            },
+          });
+          return;
+        }
       }
 
       await ack();
@@ -819,9 +501,8 @@ export function onboardSiteModal(lambdaContext) {
           };
           site.setDeliveryConfig(updatedDeliveryConfig);
 
-          // Update authoring type if crosswalk was selected
-          if (authoringType === 'crosswalk') {
-            site.setAuthoringType('crosswalk');
+          if (authoringType) {
+            site.setAuthoringType(authoringType);
           }
 
           await site.save();
