+++ src/support/utils.js
@@ -9,10 +9,18 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { Site as SiteModel } from '@adobe/spacecat-shared-data-access';
+import { Site as SiteModel, Organization as OrganizationModel } from '@adobe/spacecat-shared-data-access';
+import { Config } from '@adobe/spacecat-shared-data-access/src/models/site/config.js';
 import { ImsPromiseClient } from '@adobe/spacecat-shared-ims-client';
 import URI from 'urijs';
-import { hasText, tracingFetch as fetch } from '@adobe/spacecat-shared-utils';
+import {
+  hasText,
+  tracingFetch as fetch,
+  isValidUrl,
+  isObject,
+  resolveCanonicalUrl,
+} from '@adobe/spacecat-shared-utils';
+import { SFNClient, StartExecutionCommand } from '@aws-sdk/client-sfn';
 import {
   STATUS_BAD_REQUEST,
 } from '../utils/constants.js';
@@ -444,3 +452,401 @@ export function buildS3Prefix(type, siteId, path = '') {
   const normalized = path ? `${path.replace(/^\/+/g, '').replace(/\/+$/g, '')}/` : '';
   return `${type}/${siteId}/${normalized}`;
 }
+
+/**
+ * Checks if a site is in the LA_CUSTOMERS list and returns appropriate response if it is.
+ */
+const checkLACustomerRestriction = async (
+  baseURL,
+  imsOrgID,
+  profileName,
+  slackContext,
+  env,
+  log,
+) => {
+  const { say } = slackContext;
+
+  if (env.LA_CUSTOMERS) {
+    const laCustomers = env.LA_CUSTOMERS.split(',').map((url) => url.trim());
+    const isLACustomer = laCustomers.some(
+      (url) => baseURL.toLowerCase().endsWith(url.toLowerCase()),
+    );
+
+    if (isLACustomer) {
+      const message = `:warning: Cannot onboard site ${baseURL} - it's already onboarded and live!`;
+      log.warn(message);
+      await say(message);
+      return {
+        site: baseURL,
+        imsOrgId: imsOrgID,
+        profile: profileName,
+        errors: 'Site is a Live customer',
+        status: 'Failed',
+        existingSite: 'Yes',
+      };
+    }
+  }
+
+  return null;
+};
+
+/**
+ * Shared onboarding function used by both modal and command implementations.
+ *
+ * @param {string} baseURLInput - The site URL input
+ * @param {string} imsOrganizationID - The IMS Organization ID
+ * @param {Object} configuration - The configuration object
+ * @param {Object} profile - The loaded profile configuration object
+ * @param {number} workflowWaitTime - Workflow wait time in seconds
+ * @param {Object} slackContext - Slack context object with say function
+ * @param {Object} context - Lambda context containing dataAccess, log, etc.
+ * @param {Object} additionalParams - Additional parameters
+ * @param {Object} options - Additional options
+ * @param {Function} options.urlProcessor - Function to process the URL
+ *                                          (e.g., extractURLFromSlackInput)
+ * @param {string} options.profileName - The profile name for logging and reporting
+ * @returns {Promise<Object>} Report line object
+ */
+export const onboardSingleSite = async (
+  baseURLInput,
+  imsOrganizationID,
+  configuration,
+  profile,
+  workflowWaitTime,
+  slackContext,
+  context,
+  additionalParams = {},
+  options = {},
+) => {
+  const { say } = slackContext;
+  const {
+    dataAccess, log, imsClient, env,
+  } = context;
+  const { Site, Organization } = dataAccess;
+  const sfnClient = new SFNClient();
+
+  // Process URL - allow customization for different input formats
+  const baseURL = options.urlProcessor ? options.urlProcessor(baseURLInput) : baseURLInput.trim();
+  const imsOrgID = imsOrganizationID || env.DEMO_IMS_ORG;
+
+  // Extract profile name for logging and reporting (assume it's passed in options)
+  const profileName = options.profileName || 'unknown';
+
+  // Check if site is in LA_CUSTOMERS list
+  const laCustomerCheck = await checkLACustomerRestriction(
+    baseURL,
+    imsOrgID,
+    profileName,
+    slackContext,
+    env,
+    log,
+  );
+  if (laCustomerCheck) {
+    return laCustomerCheck;
+  }
+
+  log.info(`Starting ${profileName} environment setup for site ${baseURL}`);
+  await say(`:gear: Starting ${profileName} environment setup for site ${baseURL}`);
+  await say(':key: Please make sure you have access to the AEM Shared Production Demo environment. Request access here: https://demo.adobe.com/demos/internal/AemSharedProdEnv.html');
+
+  const reportLine = {
+    site: baseURL,
+    imsOrgId: imsOrgID,
+    spacecatOrgId: '',
+    siteId: '',
+    profile: profileName,
+    deliveryType: '',
+    authoringType: additionalParams.authoringType || '',
+    imports: '',
+    audits: '',
+    errors: '',
+    status: 'Success',
+    existingSite: 'No',
+  };
+
+  try {
+    if (!isValidUrl(baseURL)) {
+      reportLine.errors = 'Invalid site base URL';
+      reportLine.status = 'Failed';
+      return reportLine;
+    }
+
+    if (!OrganizationModel.IMS_ORG_ID_REGEX.test(imsOrgID)) {
+      reportLine.errors = 'Invalid IMS Org ID';
+      reportLine.status = 'Failed';
+      return reportLine;
+    }
+
+    // check if the organization with IMS Org ID already exists; create if it doesn't
+    let organization = await Organization.findByImsOrgId(imsOrgID);
+    if (!organization) {
+      let imsOrgDetails;
+      try {
+        imsOrgDetails = await imsClient.getImsOrganizationDetails(imsOrgID);
+        log.info(`IMS Org Details: ${imsOrgDetails}`);
+      } catch (error) {
+        log.error(`Error retrieving IMS Org details: ${error.message}`);
+        reportLine.errors = `Error retrieving IMS org with the ID *${imsOrgID}*.`;
+        reportLine.status = 'Failed';
+        return reportLine;
+      }
+
+      if (!imsOrgDetails) {
+        reportLine.errors = `Could not find details of IMS org with the ID *${imsOrgID}*.`;
+        reportLine.status = 'Failed';
+        return reportLine;
+      }
+
+      organization = await Organization.create({
+        name: imsOrgDetails.orgName,
+        imsOrgId: imsOrgID,
+      });
+
+      const message = `:white_check_mark: A new organization has been created. Organization ID: ${organization.getId()} Organization name: ${organization.getName()} IMS Org ID: ${imsOrgID}.`;
+      await say(message);
+      log.info(message);
+    }
+
+    const organizationId = organization.getId();
+    log.info(`Organization ${organizationId} was successfully retrieved or created`);
+    reportLine.spacecatOrgId = organizationId;
+
+    let site = await Site.findByBaseURL(baseURL);
+    if (site) {
+      reportLine.existingSite = 'Yes';
+      reportLine.deliveryType = site.getDeliveryType();
+      log.info(`Site ${baseURL} already exists. Site ID: ${site.getId()}, Delivery Type: ${reportLine.deliveryType}`);
+
+      const siteOrgId = site.getOrganizationId();
+      if (siteOrgId !== organizationId) {
+        log.info(`:warning: :alert: Site ${baseURL} organization ID mismatch. Run below slack command to update site organization to ${organizationId}`);
+        log.info(`:fire: @spacecat set imsorg ${baseURL} ${organizationId}`);
+      }
+    } else {
+      log.info(`Site ${baseURL} doesn't exist. Finding delivery type...`);
+
+      // Use forced delivery type if provided, otherwise detect it
+      let deliveryType;
+      if (additionalParams.deliveryType
+        && Object.values(SiteModel.DELIVERY_TYPES).includes(additionalParams.deliveryType)) {
+        deliveryType = additionalParams.deliveryType;
+        log.info(`Using forced delivery type for site ${baseURL}: ${deliveryType}`);
+      } else {
+        deliveryType = await findDeliveryType(baseURL);
+        log.info(`Found delivery type for site ${baseURL}: ${deliveryType}`);
+      }
+
+      reportLine.deliveryType = deliveryType;
+      const isLive = deliveryType === SiteModel.DELIVERY_TYPES.AEM_EDGE;
+
+      const siteCreateParams = {
+        baseURL,
+        deliveryType,
+        isLive,
+        organizationId,
+      };
+
+      // Add authoring type if provided
+      if (additionalParams.authoringType
+        && Object.values(SiteModel.AUTHORING_TYPES || {})
+          .includes(additionalParams.authoringType)) {
+        siteCreateParams.authoringType = additionalParams.authoringType;
+        log.info(`Setting authoring type for site ${baseURL}: ${additionalParams.authoringType}`);
+      }
+
+      try {
+        site = await Site.create(siteCreateParams);
+      } catch (error) {
+        log.error(`Error creating site: ${error.message}`);
+        reportLine.errors = error.message;
+        reportLine.status = 'Failed';
+        return reportLine;
+      }
+    }
+
+    const siteID = site.getId();
+    log.info(`Site ${baseURL} was successfully retrieved or created. Site ID: ${siteID}`);
+    reportLine.siteId = siteID;
+
+    log.info(`Profile ${profileName} was successfully loaded`);
+
+    if (!isObject(profile)) {
+      const error = `Profile "${profileName}" not found or invalid.`;
+      log.error(error);
+      reportLine.errors = error;
+      reportLine.status = 'Failed';
+      return reportLine;
+    }
+
+    if (!isObject(profile?.audits)) {
+      const error = `Profile "${profileName}" does not have a valid audits section.`;
+      log.error(error);
+      reportLine.errors = error;
+      reportLine.status = 'Failed';
+      return reportLine;
+    }
+
+    if (!isObject(profile?.imports)) {
+      const error = `Profile "${profileName}" does not have a valid imports section.`;
+      log.error(error);
+      reportLine.errors = error;
+      reportLine.status = 'Failed';
+      return reportLine;
+    }
+
+    const importTypes = Object.keys(profile.imports);
+    reportLine.imports = importTypes.join(', ');
+    const siteConfig = site.getConfig();
+    for (const importType of importTypes) {
+      siteConfig.enableImport(importType);
+    }
+
+    log.info(`Enabled the following imports for ${siteID}: ${reportLine.imports}`);
+
+    // Resolve canonical URL for the site from the base URL
+    const resolvedUrl = await resolveCanonicalUrl(baseURL);
+    const { pathname: baseUrlPathName } = new URL(baseURL);
+    const { pathname: resolvedUrlPathName, origin: resolvedUrlOrigin } = new URL(resolvedUrl);
+
+    log.info(`Base url: ${baseURL} -> Resolved url: ${resolvedUrl} for site ${siteID}`);
+
+    // Update the fetch configuration only if the pathname is different from the resolved URL
+    if (baseUrlPathName !== resolvedUrlPathName) {
+      siteConfig.updateFetchConfig({
+        overrideBaseURL: resolvedUrlOrigin,
+      });
+    }
+
+    site.setConfig(Config.toDynamoItem(siteConfig));
+    try {
+      await site.save();
+    } catch (error) {
+      log.error(error);
+      reportLine.errors = error.message;
+      reportLine.status = 'Failed';
+      return reportLine;
+    }
+
+    log.info(`Site config successfully saved for site ${siteID}`);
+
+    for (const importType of importTypes) {
+      /* eslint-disable no-await-in-loop */
+      await triggerImportRun(
+        configuration,
+        importType,
+        siteID,
+        profile.imports[importType].startDate,
+        profile.imports[importType].endDate,
+        slackContext,
+        context,
+      );
+    }
+
+    log.info(`Triggered the following imports for site ${siteID}: ${reportLine.imports}`);
+
+    const auditTypes = Object.keys(profile.audits);
+
+    auditTypes.forEach((auditType) => {
+      configuration.enableHandlerForSite(auditType, site);
+    });
+
+    reportLine.audits = auditTypes.join(', ');
+    log.info(`Enabled the following audits for site ${siteID}: ${reportLine.audits}`);
+
+    await say(`:white_check_mark: *Enabled imports*: ${reportLine.imports} *and audits*: ${reportLine.audits}`);
+
+    // trigger audit runs
+    log.info(`Starting audits for site ${baseURL}. Audit list: ${auditTypes}`);
+    await say(`:gear: Starting audits: ${auditTypes}`);
+    for (const auditType of auditTypes) {
+      /* eslint-disable no-await-in-loop */
+      if (!configuration.isHandlerEnabledForSite(auditType, site)) {
+        await say(`:x: Will not audit site '${baseURL}' because audits of type '${auditType}' are disabled for this site.`);
+      } else {
+        await triggerAuditForSite(
+          site,
+          auditType,
+          slackContext,
+          context,
+        );
+      }
+    }
+
+    // Opportunity status job
+    const opportunityStatusJob = {
+      type: 'opportunity-status-processor',
+      siteId: siteID,
+      siteUrl: baseURL,
+      imsOrgId: imsOrgID,
+      organizationId,
+      taskContext: {
+        auditTypes,
+        slackContext: {
+          channelId: slackContext.channelId,
+          threadTs: slackContext.threadTs,
+        },
+      },
+    };
+
+    // Disable imports and audits job
+    const disableImportAndAuditJob = {
+      type: 'disable-import-audit-processor',
+      siteId: siteID,
+      siteUrl: baseURL,
+      imsOrgId: imsOrgID,
+      organizationId,
+      taskContext: {
+        importTypes,
+        auditTypes,
+        slackContext: {
+          channelId: slackContext.channelId,
+          threadTs: slackContext.threadTs,
+        },
+      },
+    };
+
+    // Demo URL job
+    const demoURLJob = {
+      type: 'demo-url-processor',
+      siteId: siteID,
+      siteUrl: baseURL,
+      imsOrgId: imsOrgID,
+      organizationId,
+      taskContext: {
+        experienceUrl: env.EXPERIENCE_URL || 'https://experience.adobe.com',
+        slackContext: {
+          channelId: slackContext.channelId,
+          threadTs: slackContext.threadTs,
+        },
+      },
+    };
+
+    log.info(`Opportunity status job: ${JSON.stringify(opportunityStatusJob)}`);
+    log.info(`Disable import and audit job: ${JSON.stringify(disableImportAndAuditJob)}`);
+    log.info(`Demo URL job: ${JSON.stringify(demoURLJob)}`);
+
+    // Prepare and start step function workflow with the necessary parameters
+    const workflowInput = {
+      opportunityStatusJob,
+      disableImportAndAuditJob,
+      demoURLJob,
+      workflowWaitTime: workflowWaitTime || env.WORKFLOW_WAIT_TIME_IN_SECONDS,
+    };
+
+    const workflowName = `onboard-${baseURL.replace(/[^a-zA-Z0-9]/g, '-')}-${Date.now()}`;
+
+    const startCommand = new StartExecutionCommand({
+      stateMachineArn: env.ONBOARD_WORKFLOW_STATE_MACHINE_ARN,
+      input: JSON.stringify(workflowInput),
+      name: workflowName,
+    });
+    await sfnClient.send(startCommand);
+  } catch (error) {
+    log.error(error);
+    reportLine.errors = error.message;
+    reportLine.status = 'Failed';
+    throw error;
+  }
+
+  return reportLine;
+};
